Listando e cadastrando contas a receber

+++ app/controllers/api/PerfilController.php
@@ -44,38 +44,60 @@ class PerfilController{
 
         $data = $validated['issues'];
         
-        $uploadImage = (new ImageUploader($req->file('foto')->data(), ['jpg', 'jpeg', 'png', 'gif', 'webp'], 2048))->uploadImage();
+        if($validatedImage['error'] == false && $validatedImage['issues']['foto'] != null){
 
-        if($uploadImage['success']){
-
-            $data['id'] = App::authSession()->get()->id;
-            $data['foto'] = $uploadImage['filename'];
+            $uploadImage = (new ImageUploader($req->file('foto')->data(), ['jpg', 'jpeg', 'png', 'gif', 'webp'], 2048))->uploadImage();
             
+            if($uploadImage['success']){
+                
+                $data['foto'] = $uploadImage['filename'];
+                $data['id'] = App::authSession()->get()->id;
+                
+                $updated = $this->perfilService->patch('id', $data);
+                
+                if($updated){
+                    
+                    $_SESSION[SESSION_LOGIN]->foto = $data['foto'];
+                    
+                    return $res->json([
+                        'error' => false,
+                        'message' => 'Perfil atualizado com sucesso.',
+                    ]);
+                    
+                }else{
+                    unlink(UPLOAD_DIR . $uploadImage['filename']);
+                    return $res->json([
+                        'error' => true,
+                        'message' => 'O perfil não foi atualizado.',
+                    ]);
+                }
+            }
+
+        }else{
+            $data['id'] = App::authSession()->get()->id;
+                
             $updated = $this->perfilService->patch('id', $data);
-
+            
             if($updated){
-
-                $_SESSION[SESSION_LOGIN]->foto = $data['foto'];
-
+                                
                 return $res->json([
                     'error' => false,
                     'message' => 'Perfil atualizado com sucesso.',
                 ]);
                 
             }else{
-                unlink(UPLOAD_DIR . $uploadImage['filename']);
                 return $res->json([
                     'error' => true,
                     'message' => 'O perfil não foi atualizado.',
                 ]);
             }
-
-        }else{
-            return $res->json([
-                'error' => true,
-                'message' => 'Não foi possível realizar o upload de imagem por motivo diverso. ',
-            ]);
+            
         }
+
+        return $res->json([
+                'error' => true,
+                'message' => 'Ocorreu um erro na hora realizar a atualização do perfil.',
+            ]);
         
     }
